src/lp_server/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.8.2
boto3==1.17.24
botocore==1.20.24
celery==5.2.3
certifi==2023.07.22
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.24
django-admin-rangefilter==0.8.1
django-bleach==0.6.1
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-countries==7.3.2
django-debug-toolbar==3.2.1
django-jalali==3.3.0
django-js-asset==1.2.2
django-modeltranslation==0.15.1
django-redis==5.0.0
django-storages==1.9.1
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
django-taggit==2.0
docutils==0.15.2
gunicorn==20.1.0
idna==2.9
itypes==1.1.0
jdatetime==3.6.2
jmespath==0.9.5
MarkupSafe==1.1.1
model-bakery==1.1.0
openapi-codec==1.3.2
outline-api
prometheus-client==0.13.0
psycopg2==2.8.4
python-dateutil==2.8.1
python-magic==0.4.27
requests==2.31.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
strawberry-graphql==0.209.2
strawberry-django-auth==0.375.0
strawberry-graphql-django==0.16.0
uritemplate==3.0.1
wagtail==4.1.9
wagtailmedia==0.12
webencodings==0.5.1
django-solo==2.0.0
